Rewrote functions to handle dynamic acquisitions.
Assignment of images/slices to dynamic volumes is more intelligent and no
longer switches to a new volume based on changes in the sequence of slice
locations. New method tracks if a given slice (described by its IPP-IOP
vector) is already present in the set of slices for the given volume - if
so then a new volume is started. Also missing and extra slices are handled
in a more intelligent manner.